backend-raseed/requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.116.0
uvicorn==0.35.0
python-multipart==0.0.20
aiofiles==24.1.0

# Google Cloud and AI services
google-generativeai==0.8.3
google-cloud-aiplatform==1.48.0
google-auth-oauthlib==1.2.0
google-auth==2.28.1

# Data processing and validation
pydantic==2.11.7
requests==2.32.3

# Image processing
opencv-python==4.9.0.80
Pillow==10.4.0
numpy==1.26.4

# JWT and authentication
PyJWT==2.8.0

# Environment and configuration
python-dotenv==1.0.1

# Utilities
colorama==0.4.6

# Additional dependencies found in the codebase
vertexai==1.48.0 